requirements.txt
torchquantum
torchtext
scikit-learn
datasets
qiskit<1.0
qiskit-aer==0.13.3
qiskit-ibm-provider==0.10.0
qiskit-ibm-runtime==0.20.0
qiskit-terra